fix motion detection for larger framesizes

+++ src/ImageProcessing/MotionDetection.h
@@ -155,7 +155,7 @@ namespace Eloquent {
             /**
              * Cache number of changes
              */
-            uint8_t _changes;
+            uint16_t _changes;
         };
     }
 }
